+++ TodoMate/Service/AuthService.swift
@@ -1,0 +1,55 @@
+//
+//  AuthManager.swift
+//  TodoMate
+//
+//  Created by hs on 1/19/25.
+//
+
+import SwiftUI
+
+protocol AuthServiceType {
+    func signIn() async -> User?
+    func signOut() async
+}
+
+final class AuthService: AuthServiceType {
+    private let userService: UserServiceType
+    private let googleSignInService: GoogleSignInServiceType
+    
+    init(userService: UserServiceType, googleSignInService: GoogleSignInServiceType) {
+        self.userService = userService
+        self.googleSignInService = googleSignInService
+    }
+    
+    func signIn() async -> User? {
+        guard let gUser = try? await googleSignInService.signIn() else {
+            print("[AuthService] Failed to sign in")
+            return nil
+        }
+        
+        guard let existingUser = await userService.fetch(uid: gUser.uid) else {
+            let newUser = User(uid: gUser.uid, nickname: gUser.name ?? "Unknown")
+            
+            await userService.update(newUser)
+            print("[AuthService] Created new user: \(newUser.uid)")
+            return newUser
+        }
+        
+        print("[AuthService] User already exists with uid: \(existingUser.uid)")
+        return existingUser
+    }
+    
+    func signOut() async {
+        googleSignInService.signOut()
+    }
+}
+
+final class StubAuthService: AuthServiceType {
+    func signIn() async -> User? {
+        return User.stub[0]
+    }
+    
+    func signOut() async {}
+}
+    
+
